More 64 bitflag fixes. Made dynamic dnxClient caching work.
git-svn-id: https://svn.forschooner.net/code/dnx/trunk/dnx@302 2be5450d-c085-40c6-83eb-9dc9c13f2dba

+++ server/dnxRegistrar.h
@@ -84,7 +84,7 @@ void dnxRegistrarDestroy(DnxRegistrar * reg);
 *
 * @return Affinity object on success, NULL on failure.
 */
-DnxAffinityList * addDnxAffinity(DnxAffinityList *p, char * name, unsigned int flag);
+DnxAffinityList * addDnxAffinity(DnxAffinityList *p, char * name, unsigned long long flag);
 
 #endif   /* _DNXREGISTRAR_H_ */
